src/ozon_parser/requirements.txt
Django==4.1.12
djangorestframework==3.14.0
drf-yasg==1.21.7
beautifulsoup4==4.12.2
django-adminlte-ui==2.1.1
celery[redis]==5.3.4
aiogram==3.1.1
lxml==4.9.3
mysqlclient==2.2.0
requests==2.31.0
python-dotenv==1.0.0